Re-add class-route file
